+++ app/app_test.go
@@ -22,9 +22,9 @@ import (
 	babylonApp "github.com/babylonlabs-io/babylon/v3/app"
 	testsigner "github.com/babylonlabs-io/babylon/v3/testutil/signer"
 	checkpointingtypes "github.com/babylonlabs-io/babylon/v3/x/checkpointing/types"
+	epochingtypes "github.com/babylonlabs-io/babylon/v3/x/epoching/types"
 	incentivetypes "github.com/babylonlabs-io/babylon/v3/x/incentive/types"
 	minttypes "github.com/babylonlabs-io/babylon/v3/x/mint/types"
-
 	"github.com/stretchr/testify/require"
 )
 
@@ -32,6 +32,7 @@ var (
 	expectedMaccPerms = map[string][]string{
 		authtypes.FeeCollectorName:                  nil, // fee collector account
 		distrtypes.ModuleName:                       nil,
+		epochingtypes.DelegatePoolModuleName:        nil,
 		minttypes.ModuleName:                        {authtypes.Minter},
 		stktypes.BondedPoolName:                     {authtypes.Burner, authtypes.Staking},
 		stktypes.NotBondedPoolName:                  {authtypes.Burner, authtypes.Staking},