+++ x/epoching/types/params.pb.go
@@ -27,6 +27,10 @@ const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 type Params struct {
 	// epoch_interval is the number of consecutive blocks to form an epoch
 	EpochInterval uint64 `protobuf:"varint,1,opt,name=epoch_interval,json=epochInterval,proto3" json:"epoch_interval,omitempty" yaml:"epoch_interval"`
+	// execute_gas defines raw gas for different executions.
+	ExecuteGas ExecuteGas `protobuf:"bytes,2,opt,name=execute_gas,json=executeGas,proto3" json:"execute_gas" yaml:"execute_gas"`
+	// minimum_amount is a minimum amount for staking message cancel_unbonding_delegation
+	MinAmount uint64 `protobuf:"varint,3,opt,name=min_amount,json=minAmount,proto3" json:"min_amount,omitempty" yaml:"min_amount"`
 }
 
 func (m *Params) Reset()         { *m = Params{} }
@@ -69,27 +73,144 @@ func (m *Params) GetEpochInterval() uint64 {
 	return 0
 }
 
+func (m *Params) GetExecuteGas() ExecuteGas {
+	if m != nil {
+		return m.ExecuteGas
+	}
+	return ExecuteGas{}
+}
+
+func (m *Params) GetMinAmount() uint64 {
+	if m != nil {
+		return m.MinAmount
+	}
+	return 0
+}
+
+// ExecuteGas defines the raw gas for the enqueued message execution.
+type ExecuteGas struct {
+	Delegate                  uint64 `protobuf:"varint,1,opt,name=delegate,proto3" json:"delegate,omitempty" yaml:"delegate"`
+	Undelegate                uint64 `protobuf:"varint,2,opt,name=undelegate,proto3" json:"undelegate,omitempty" yaml:"undelegate"`
+	BeginRedelegate           uint64 `protobuf:"varint,3,opt,name=begin_redelegate,json=beginRedelegate,proto3" json:"begin_redelegate,omitempty" yaml:"begin_redelegate"`
+	CancelUnbondingDelegation uint64 `protobuf:"varint,4,opt,name=cancel_unbonding_delegation,json=cancelUnbondingDelegation,proto3" json:"cancel_unbonding_delegation,omitempty" yaml:"cancel_unbonding_delegation"`
+	EditValidator             uint64 `protobuf:"varint,5,opt,name=edit_validator,json=editValidator,proto3" json:"edit_validator,omitempty" yaml:"edit_validator"`
+	CreateValidator           uint64 `protobuf:"varint,6,opt,name=create_validator,json=createValidator,proto3" json:"create_validator,omitempty" yaml:"create_validator"`
+}
+
+func (m *ExecuteGas) Reset()         { *m = ExecuteGas{} }
+func (m *ExecuteGas) String() string { return proto.CompactTextString(m) }
+func (*ExecuteGas) ProtoMessage()    {}
+func (*ExecuteGas) Descriptor() ([]byte, []int) {
+	return fileDescriptor_c9e38cfe55335900, []int{1}
+}
+func (m *ExecuteGas) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *ExecuteGas) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_ExecuteGas.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *ExecuteGas) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ExecuteGas.Merge(m, src)
+}
+func (m *ExecuteGas) XXX_Size() int {
+	return m.Size()
+}
+func (m *ExecuteGas) XXX_DiscardUnknown() {
+	xxx_messageInfo_ExecuteGas.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_ExecuteGas proto.InternalMessageInfo
+
+func (m *ExecuteGas) GetDelegate() uint64 {
+	if m != nil {
+		return m.Delegate
+	}
+	return 0
+}
+
+func (m *ExecuteGas) GetUndelegate() uint64 {
+	if m != nil {
+		return m.Undelegate
+	}
+	return 0
+}
+
+func (m *ExecuteGas) GetBeginRedelegate() uint64 {
+	if m != nil {
+		return m.BeginRedelegate
+	}
+	return 0
+}
+
+func (m *ExecuteGas) GetCancelUnbondingDelegation() uint64 {
+	if m != nil {
+		return m.CancelUnbondingDelegation
+	}
+	return 0
+}
+
+func (m *ExecuteGas) GetEditValidator() uint64 {
+	if m != nil {
+		return m.EditValidator
+	}
+	return 0
+}
+
+func (m *ExecuteGas) GetCreateValidator() uint64 {
+	if m != nil {
+		return m.CreateValidator
+	}
+	return 0
+}
+
 func init() {
 	proto.RegisterType((*Params)(nil), "babylon.epoching.v1.Params")
+	proto.RegisterType((*ExecuteGas)(nil), "babylon.epoching.v1.ExecuteGas")
 }
 
 func init() { proto.RegisterFile("babylon/epoching/v1/params.proto", fileDescriptor_c9e38cfe55335900) }
 
 var fileDescriptor_c9e38cfe55335900 = []byte{
-	// 206 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0x52, 0x48, 0x4a, 0x4c, 0xaa,
-	0xcc, 0xc9, 0xcf, 0xd3, 0x4f, 0x2d, 0xc8, 0x4f, 0xce, 0xc8, 0xcc, 0x4b, 0xd7, 0x2f, 0x33, 0xd4,
-	0x2f, 0x48, 0x2c, 0x4a, 0xcc, 0x2d, 0xd6, 0x2b, 0x28, 0xca, 0x2f, 0xc9, 0x17, 0x12, 0x86, 0xaa,
-	0xd0, 0x83, 0xa9, 0xd0, 0x2b, 0x33, 0x94, 0x12, 0x49, 0xcf, 0x4f, 0xcf, 0x07, 0xcb, 0xeb, 0x83,
-	0x58, 0x10, 0xa5, 0x4a, 0x01, 0x5c, 0x6c, 0x01, 0x60, 0xad, 0x42, 0x0e, 0x5c, 0x7c, 0x60, 0xe5,
-	0xf1, 0x99, 0x79, 0x25, 0xa9, 0x45, 0x65, 0x89, 0x39, 0x12, 0x8c, 0x0a, 0x8c, 0x1a, 0x2c, 0x4e,
-	0x92, 0x9f, 0xee, 0xc9, 0x8b, 0x56, 0x26, 0xe6, 0xe6, 0x58, 0x29, 0xa1, 0xca, 0x2b, 0x05, 0xf1,
-	0x82, 0x05, 0x3c, 0xa1, 0x7c, 0x2b, 0x96, 0x17, 0x0b, 0xe4, 0x19, 0x9d, 0xfc, 0x4f, 0x3c, 0x92,
-	0x63, 0xbc, 0xf0, 0x48, 0x8e, 0xf1, 0xc1, 0x23, 0x39, 0xc6, 0x09, 0x8f, 0xe5, 0x18, 0x2e, 0x3c,
-	0x96, 0x63, 0xb8, 0xf1, 0x58, 0x8e, 0x21, 0xca, 0x34, 0x3d, 0xb3, 0x24, 0xa3, 0x34, 0x49, 0x2f,
-	0x39, 0x3f, 0x57, 0x1f, 0xea, 0xc2, 0x9c, 0xc4, 0xa4, 0x62, 0xdd, 0xcc, 0x7c, 0x18, 0x57, 0xbf,
-	0xcc, 0x58, 0xbf, 0x02, 0xe1, 0xaf, 0x92, 0xca, 0x82, 0xd4, 0xe2, 0x24, 0x36, 0xb0, 0x4b, 0x8d,
-	0x01, 0x01, 0x00, 0x00, 0xff, 0xff, 0x75, 0x5e, 0x28, 0xdc, 0xf8, 0x00, 0x00, 0x00,
+	// 469 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x7c, 0x93, 0xcd, 0x6e, 0xd4, 0x30,
+	0x10, 0xc7, 0x37, 0xed, 0xb2, 0x02, 0x57, 0xb0, 0xe0, 0x52, 0x91, 0xb6, 0x52, 0xbc, 0xf2, 0x01,
+	0xf5, 0x42, 0xa2, 0x52, 0x7a, 0xe9, 0x09, 0x56, 0x7c, 0x88, 0x13, 0xc8, 0x12, 0x1c, 0x10, 0x52,
+	0xe4, 0x24, 0x26, 0xb5, 0x94, 0xd8, 0xab, 0xc4, 0x89, 0xba, 0x6f, 0xc1, 0x23, 0xf0, 0x10, 0x3c,
+	0x44, 0x8f, 0x3d, 0x72, 0x8a, 0xd0, 0x2e, 0x07, 0xce, 0x79, 0x02, 0xb4, 0x76, 0x3e, 0xda, 0x80,
+	0xb8, 0x79, 0xe6, 0xff, 0x9b, 0xff, 0xcc, 0x28, 0x13, 0x30, 0x0b, 0x68, 0xb0, 0x4c, 0xa4, 0xf0,
+	0xd8, 0x42, 0x86, 0xe7, 0x5c, 0xc4, 0x5e, 0x79, 0xec, 0x2d, 0x68, 0x46, 0xd3, 0xdc, 0x5d, 0x64,
+	0x52, 0x49, 0xb8, 0xdb, 0x10, 0x6e, 0x4b, 0xb8, 0xe5, 0xf1, 0xc1, 0xc3, 0x58, 0xc6, 0x52, 0xeb,
+	0xde, 0xe6, 0x65, 0x50, 0xfc, 0xcb, 0x02, 0x93, 0xf7, 0xba, 0x16, 0x3e, 0x07, 0xf7, 0x34, 0xef,
+	0x73, 0xa1, 0x58, 0x56, 0xd2, 0xc4, 0xb6, 0x66, 0xd6, 0xd1, 0x78, 0xbe, 0x5f, 0x57, 0x68, 0x6f,
+	0x49, 0xd3, 0xe4, 0x0c, 0xdf, 0xd4, 0x31, 0xb9, 0xab, 0x13, 0x6f, 0x9b, 0x18, 0x7e, 0x06, 0x3b,
+	0xec, 0x82, 0x85, 0x85, 0x62, 0x7e, 0x4c, 0x73, 0x7b, 0x6b, 0x66, 0x1d, 0xed, 0x3c, 0x45, 0xee,
+	0x3f, 0xa6, 0x71, 0x5f, 0x19, 0xee, 0x0d, 0xcd, 0xe7, 0x07, 0x97, 0x15, 0x1a, 0xd5, 0x15, 0x82,
+	0x4d, 0x8f, 0xde, 0x01, 0x13, 0xc0, 0x3a, 0x0e, 0x3e, 0x03, 0x20, 0xe5, 0xc2, 0xa7, 0xa9, 0x2c,
+	0x84, 0xb2, 0xb7, 0xf5, 0x6c, 0x7b, 0x75, 0x85, 0x1e, 0x98, 0xba, 0x5e, 0xc3, 0xe4, 0x4e, 0xca,
+	0xc5, 0x0b, 0xfd, 0x3e, 0x1b, 0xff, 0xfe, 0x86, 0x2c, 0xfc, 0x7d, 0x1b, 0x80, 0xbe, 0x25, 0xf4,
+	0xc0, 0xed, 0x88, 0x25, 0x2c, 0xa6, 0x8a, 0x35, 0x4b, 0xee, 0xd6, 0x15, 0x9a, 0x1a, 0xa3, 0x56,
+	0xc1, 0xa4, 0x83, 0xe0, 0x29, 0x00, 0x85, 0xe8, 0x4a, 0xb6, 0x86, 0xbd, 0x7b, 0x0d, 0x93, 0x6b,
+	0x20, 0x7c, 0x0d, 0xee, 0x07, 0x2c, 0xe6, 0xc2, 0xcf, 0x58, 0x57, 0x6c, 0x06, 0x3f, 0xac, 0x2b,
+	0xf4, 0xc8, 0x14, 0x0f, 0x09, 0x4c, 0xa6, 0x3a, 0x45, 0xba, 0x0c, 0xfc, 0x02, 0x0e, 0x43, 0x2a,
+	0x42, 0x96, 0xf8, 0x85, 0x08, 0xa4, 0x88, 0xb8, 0x88, 0xfd, 0x46, 0xe4, 0x52, 0xd8, 0x63, 0x6d,
+	0xf9, 0xb8, 0xae, 0x10, 0x36, 0x96, 0xff, 0x81, 0x31, 0xd9, 0x37, 0xea, 0x87, 0x56, 0x7c, 0xd9,
+	0x69, 0xfa, 0x04, 0x22, 0xae, 0xfc, 0x92, 0x26, 0x3c, 0xa2, 0x4a, 0x66, 0xf6, 0xad, 0xbf, 0x4e,
+	0xe0, 0x86, 0xbe, 0x39, 0x81, 0x88, 0xab, 0x8f, 0x6d, 0xbc, 0xd9, 0x38, 0xcc, 0x18, 0x55, 0xec,
+	0x9a, 0xc7, 0x64, 0xb8, 0xf1, 0x90, 0xc0, 0x64, 0x6a, 0x52, 0x9d, 0x8f, 0xf9, 0x6c, 0xf3, 0x77,
+	0x97, 0x2b, 0xc7, 0xba, 0x5a, 0x39, 0xd6, 0xcf, 0x95, 0x63, 0x7d, 0x5d, 0x3b, 0xa3, 0xab, 0xb5,
+	0x33, 0xfa, 0xb1, 0x76, 0x46, 0x9f, 0x4e, 0x63, 0xae, 0xce, 0x8b, 0xc0, 0x0d, 0x65, 0xea, 0x35,
+	0xf7, 0x95, 0xd0, 0x20, 0x7f, 0xc2, 0x65, 0x1b, 0x7a, 0xe5, 0x89, 0x77, 0xd1, 0xff, 0x23, 0x6a,
+	0xb9, 0x60, 0x79, 0x30, 0xd1, 0x57, 0x7f, 0xf2, 0x27, 0x00, 0x00, 0xff, 0xff, 0xca, 0x6d, 0xf3,
+	0xdc, 0x44, 0x03, 0x00, 0x00,
 }
 
 func (this *Params) Equal(that interface{}) bool {
@@ -114,6 +235,51 @@ func (this *Params) Equal(that interface{}) bool {
 	if this.EpochInterval != that1.EpochInterval {
 		return false
 	}
+	if !this.ExecuteGas.Equal(&that1.ExecuteGas) {
+		return false
+	}
+	if this.MinAmount != that1.MinAmount {
+		return false
+	}
+	return true
+}
+func (this *ExecuteGas) Equal(that interface{}) bool {
+	if that == nil {
+		return this == nil
+	}
+
+	that1, ok := that.(*ExecuteGas)
+	if !ok {
+		that2, ok := that.(ExecuteGas)
+		if ok {
+			that1 = &that2
+		} else {
+			return false
+		}
+	}
+	if that1 == nil {
+		return this == nil
+	} else if this == nil {
+		return false
+	}
+	if this.Delegate != that1.Delegate {
+		return false
+	}
+	if this.Undelegate != that1.Undelegate {
+		return false
+	}
+	if this.BeginRedelegate != that1.BeginRedelegate {
+		return false
+	}
+	if this.CancelUnbondingDelegation != that1.CancelUnbondingDelegation {
+		return false
+	}
+	if this.EditValidator != that1.EditValidator {
+		return false
+	}
+	if this.CreateValidator != that1.CreateValidator {
+		return false
+	}
 	return true
 }
 func (m *Params) Marshal() (dAtA []byte, err error) {
@@ -136,8 +302,76 @@ func (m *Params) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
+	if m.MinAmount != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.MinAmount))
+		i--
+		dAtA[i] = 0x18
+	}
+	{
+		size, err := m.ExecuteGas.MarshalToSizedBuffer(dAtA[:i])
+		if err != nil {
+			return 0, err
+		}
+		i -= size
+		i = encodeVarintParams(dAtA, i, uint64(size))
+	}
+	i--
+	dAtA[i] = 0x12
 	if m.EpochInterval != 0 {
 		i = encodeVarintParams(dAtA, i, uint64(m.EpochInterval))
+		i--
+		dAtA[i] = 0x8
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *ExecuteGas) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *ExecuteGas) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *ExecuteGas) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.CreateValidator != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.CreateValidator))
+		i--
+		dAtA[i] = 0x30
+	}
+	if m.EditValidator != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.EditValidator))
+		i--
+		dAtA[i] = 0x28
+	}
+	if m.CancelUnbondingDelegation != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.CancelUnbondingDelegation))
+		i--
+		dAtA[i] = 0x20
+	}
+	if m.BeginRedelegate != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.BeginRedelegate))
+		i--
+		dAtA[i] = 0x18
+	}
+	if m.Undelegate != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.Undelegate))
+		i--
+		dAtA[i] = 0x10
+	}
+	if m.Delegate != 0 {
+		i = encodeVarintParams(dAtA, i, uint64(m.Delegate))
 		i--
 		dAtA[i] = 0x8
 	}
@@ -163,6 +397,38 @@ func (m *Params) Size() (n int) {
 	_ = l
 	if m.EpochInterval != 0 {
 		n += 1 + sovParams(uint64(m.EpochInterval))
+	}
+	l = m.ExecuteGas.Size()
+	n += 1 + l + sovParams(uint64(l))
+	if m.MinAmount != 0 {
+		n += 1 + sovParams(uint64(m.MinAmount))
+	}
+	return n
+}
+
+func (m *ExecuteGas) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if m.Delegate != 0 {
+		n += 1 + sovParams(uint64(m.Delegate))
+	}
+	if m.Undelegate != 0 {
+		n += 1 + sovParams(uint64(m.Undelegate))
+	}
+	if m.BeginRedelegate != 0 {
+		n += 1 + sovParams(uint64(m.BeginRedelegate))
+	}
+	if m.CancelUnbondingDelegation != 0 {
+		n += 1 + sovParams(uint64(m.CancelUnbondingDelegation))
+	}
+	if m.EditValidator != 0 {
+		n += 1 + sovParams(uint64(m.EditValidator))
+	}
+	if m.CreateValidator != 0 {
+		n += 1 + sovParams(uint64(m.CreateValidator))
 	}
 	return n
 }
@@ -217,6 +483,222 @@ func (m *Params) Unmarshal(dAtA []byte) error {
 				b := dAtA[iNdEx]
 				iNdEx++
 				m.EpochInterval |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ExecuteGas", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthParams
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthParams
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if err := m.ExecuteGas.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 3:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field MinAmount", wireType)
+			}
+			m.MinAmount = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.MinAmount |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		default:
+			iNdEx = preIndex
+			skippy, err := skipParams(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthParams
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *ExecuteGas) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowParams
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: ExecuteGas: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: ExecuteGas: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Delegate", wireType)
+			}
+			m.Delegate = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.Delegate |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 2:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Undelegate", wireType)
+			}
+			m.Undelegate = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.Undelegate |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 3:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field BeginRedelegate", wireType)
+			}
+			m.BeginRedelegate = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.BeginRedelegate |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 4:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field CancelUnbondingDelegation", wireType)
+			}
+			m.CancelUnbondingDelegation = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.CancelUnbondingDelegation |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 5:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field EditValidator", wireType)
+			}
+			m.EditValidator = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.EditValidator |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 6:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field CreateValidator", wireType)
+			}
+			m.CreateValidator = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.CreateValidator |= uint64(b&0x7F) << shift
 				if b < 0x80 {
 					break
 				}